(BSR)[API] refactor: remove unnecessary educational api functions
