+++ api/src/pcapi/routes/adage/v1/educational_institution.py
@@ -1,8 +1,6 @@
 import logging
 
-from pcapi.core.educational.api.adage import find_collective_bookings_for_adage
-from pcapi.core.educational.api.deposit import find_educational_deposit_by_institution_id_and_year
-from pcapi.core.educational.repository import find_educational_institution_by_uai_code
+from pcapi.core.educational import repository
 from pcapi.core.educational.serialization import collective_booking as collective_booking_serialize
 from pcapi.models.api_errors import ApiErrors
 from pcapi.repository.session_management import atomic
@@ -28,15 +26,15 @@ educational_institution_path = "years/<string:year_id>/educational_institution/<
 )
 @adage_api_key_required
 def get_educational_institution(year_id: str, uai_code: str) -> EducationalInstitutionResponse:
-    educational_institution = find_educational_institution_by_uai_code(uai_code)
+    educational_institution = repository.find_educational_institution_by_uai_code(uai_code)
 
     if not educational_institution:
         raise ApiErrors({"code": "EDUCATIONAL_INSTITUTION_NOT_FOUND"}, status_code=404)
 
-    collective_bookings = find_collective_bookings_for_adage(uai_code=uai_code, year_id=year_id)
+    collective_bookings = repository.find_collective_bookings_for_adage(uai_code=uai_code, year_id=year_id)
     prebookings = collective_booking_serialize.serialize_collective_bookings(collective_bookings)
 
-    educational_deposit = find_educational_deposit_by_institution_id_and_year(
+    educational_deposit = repository.find_educational_deposit_by_institution_id_and_year(
         educational_year_id=year_id, educational_institution_id=educational_institution.id
     )
